requirements.txt
Flask
opencv-python
pygame
PyOpenGL
mediapipe